querier: clamp start time for Series, LabelValues, LabelNames
Start and End times in Series, LabelNames, LabelValues calls are optional.
When they're unset Prometheus defaults them to MinTime (effectively MinInt)
and MaxTime (MaxInt).

Currently MaxQueryIntoFuture clamps the start time relative to now.
After that the start time is clamped relative to the end because of the max_label_query_length limit.

However with PR XXX we're preparing to remove that option, so we'll no longer
be clamping max time relative to now. Without clamping
the end time we end up clamping start to MaxTime, so the whole query is for
some (e.g.) 30 days in the year 292277025. Which usually returns no data.

What does this PR do

In this PR we're only clamping start time.
It's clamped to end time if that was provided or to now, otherwise.
We don't clamp end time because there's likely no data to query and
we don't have to think about ignoring samples with a future timestamp that have been ingested.
Not clamping end time is potentially a problem if there's large clock skew
between clients sending data and the query-frontend. I think this is very unlikely to cause problems.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>